fix(desktop): repair Windows compile breaks and harden SID seam
The #282/#295/#296 extraction moved code out of main.rs but left several
Windows-only symbol references unimported -- invisible on macOS, fatal on
Windows CI. Alongside the compile repair, the connect path's same-user
identity check is made structural rather than convention-enforced, and the
relay owner field's race is closed at the type-seam level.

Windows compile breaks (each compiler-verified via
`cargo check --target x86_64-pc-windows-gnu`):
- finalize_sidecar_streams was imported unconditionally but is unix-only
  (named pipes have no socket timeout to clear); gate the import #[cfg(unix)].
- Five windows test helpers (require_peer_sid, require_same_user_pipe_peer,
  sanitize_sid_for_pipe, sidecar_identity, dev_sidecar_metadata_path_in) were
  called bare in #[cfg(windows)] tests but never re-imported; add a
  #[cfg(all(windows, test))] import block.
- Two windows tests call server.read().await needing AsyncReadExt in scope,
  which the move dropped; re-add it under #[cfg(all(windows, test))].
- ClientOptions was imported under #[cfg(windows)] but used only in tests,
  tripping -D warnings in the production binary; gate it under test and keep
  NamedPipeClient (production) separate.
- The windows-only connect/handshake (windows_handshake_async,
  connect_and_handshake_dev_sidecar) and its windows-only tests now live in a
  #[cfg(windows)] mod windows_impl, so every windows-only import is gated
  automatically by the module and a future one cannot compile silently on
  macOS and fail only on Windows CI -- the trap the per-import gates above
  were set up to prevent.

SID-seam hardening (structural, not convention-enforced):
- connect_sidecar_endpoint (sync) and endpoint_holder_pid (windows) route
  through connect_sidecar_endpoint_async, the single connect+SID-check core,
  so every connect that can adopt or query a peer is gated. Closes the gap
  that let the handshake and the holder-pid diagnostic bypass
  require_same_user_pipe_peer.
- The unix endpoint_holder_pid twin now applies require_same_user_peer on its
  throwaway connection too, so the OS-agnostic "we only ever want the holder
  pid of OUR sidecar" invariant holds on macOS/Linux as well; a foreign-user
  socket squatting the dev endpoint reports no holder rather than poisoning
  the diagnostic with its pid.
- Cap the GetLengthSid allocation in token_user_sid off SECURITY_MAX_SID_SIZE
  (68) * 4, so a malformed/huge length fails closed instead of driving a
  multi-GiB vec! on the connect path off a peer-controlled pipe name.
- Cache current_user_sid (the byte form) for the process lifetime, mirroring
  sidecar_identity's OnceLock; the SID check runs every iteration of the
  ~600-iteration cold-start connect retry loop, so this drops a redundant
  OpenProcessToken+GetTokenInformation round-trip per connect.
- Tighten sidecar_ipc/mod.rs: replace `pub(crate) use unix::*`/`windows::*`
  globs with per-item re-exports; demote open_named_pipe_client/PipeConnect
  to module-private so a future connect path cannot reach past the
  SID-checking seam.

Relay owner race (closed at the access-seam level):
- Convert all wsRelay.owner reads/writes to standalone atomic ops
  (atomic.LoadUint64/StoreUint64). The field stays a plain uint64 because
  atomic.Uint64's noCopy conflicts with wsRelay's value-embedding in
  ChannelRelay/OrgEventsRelay; standalone atomic ops are safe on every
  target LeapMux builds for (macOS/Linux/Windows, all 64-bit).
- Encapsulate the atomic access behind wsRelay.ownerNow/stampOwner so the
  ~14 production + test call sites share one named seam and a future bare
  `relay.owner` read cannot silently bypass the atomic.
- Correct the field/method docs that mis-described the access: the org-events
  rejectIfSuperseded only LOADS owner (it refuses, never re-stamps), so it is
  not a re-stamp path; and the forOwner doc no longer claims the field is an
  atomic.Uint64 (it is not). Fix the stale OpenChannelRelay doc that pointed
  at a non-existent desktopConnection.relayOwner field.

Also: restore the realistic SidecarInfo payload in the multibyte-varint
roundtrip test (the move had narrowed it to a synthetic error string).

Verified locally: cargo +1.97.1 clippy --all-targets -- -D warnings clean;
55 Rust tests pass; cargo fmt clean; Windows cross-compile clean (only the
pre-existing embed_plist Tauri artifact remains); the windows_impl module
type-checks clean on the windows target (verified by stubbing the
embed_plist-blocked generate_context); golangci-lint 0 issues; go test -race
green. Windows CI is the final gate for the Windows-only Rust changes (not
runnable on this macOS host).
