+++ desktop/go/app.go
@@ -316,7 +316,7 @@ func (a *App) CloseRelayForUndeliverableEvent(emitterOwner uint64, event *deskto
 // log says, and which close event the frontend receives, so those ride as
 // parameters the way wsRelay.emitClose takes its event builder.
 //
-// The ownership gate (getRelay(...).owner == emitterOwner) is the point of the
+// The ownership gate (getRelay(...).ownerNow() == emitterOwner) is the point of the
 // emitterOwner parameter: without it, a close spawned by emitter A's read loop
 // could execute after emitter B's open superseded A and tear down B's relay,
 // forcing B to reconnect for A's fault. The gate makes the close a no-op when
@@ -327,7 +327,7 @@ func (a *App) closeUndeliverableRelay(getRelay func(*desktopConnection) *wsRelay
 	a.lifecycleMu.Lock()
 	var drain <-chan struct{}
 	if a.connection != nil {
-		if relay := getRelay(a.connection); relay != nil && relay.owner == emitterOwner {
+		if relay := getRelay(a.connection); relay != nil && relay.ownerNow() == emitterOwner {
 			drain = closeRelay()
 		}
 	}
@@ -360,13 +360,9 @@ func (a *App) SetEventSinkForRelay(sink func(owner uint64, event *desktoppb.Even
 	a.events.SetRelay(sink)
 }
 
-// emitForOwner returns a func the relay read loops call in place of a bare
-// EmitEvent, capturing the emitting relay so the owner id at emit time rides
-// alongside the event. The closure reads relay.owner at CALL time rather than
-// capturing its value, because owner is stamped AFTER newWSRelay constructs the
-// relay -- but it is stable for the relay's lifetime once stamped (a supersede
-// replaces connection.relay, not this relay's owner field), so the read loop's
-// emit always reports the id this relay was installed under.
+// emitForOwner is a thin forwarder onto eventSink.forOwner; see that method for
+// why the closure reads relay.owner at call time (via an atomic load against the
+// lifecycleMu-guarded re-stamp an adopt performs) rather than capturing its value.
 func (a *App) emitForOwner(relay *wsRelay) func(*desktoppb.Event) {
 	return a.events.forOwner(relay)
 }
